cleanup component, upgrade dependencies
